Fixed some bugs
